+++ builder/builder_numeric.go
@@ -34,7 +34,7 @@ type bigDecimalFloatBuilder struct{}
 
 var globalBigDecimalFloatBuilder = &bigDecimalFloatBuilder{}
 
-func generateBigDecimalFloatBuilder(ctx *Context) ObjectBuilder { return globalBigDecimalFloatBuilder }
+func generateBigDecimalFloatBuilder(ctx *Context) Builder { return globalBigDecimalFloatBuilder }
 func (_this *bigDecimalFloatBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *bigDecimalFloatBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
@@ -72,7 +72,7 @@ type bigFloatBuilder struct{}
 
 var globalBigFloatBuilder = &bigFloatBuilder{}
 
-func generateBigFloatBuilder(ctx *Context) ObjectBuilder { return globalBigFloatBuilder }
+func generateBigFloatBuilder(ctx *Context) Builder { return globalBigFloatBuilder }
 func (_this *bigFloatBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *bigFloatBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
@@ -110,7 +110,7 @@ type bigIntBuilder struct{}
 
 var globalBigIntBuilder = &bigIntBuilder{}
 
-func generateBigIntBuilder(ctx *Context) ObjectBuilder { return globalBigIntBuilder }
+func generateBigIntBuilder(ctx *Context) Builder { return globalBigIntBuilder }
 func (_this *bigIntBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *bigIntBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
@@ -148,7 +148,7 @@ type decimalFloatBuilder struct{}
 
 var globalDecimalFloatBuilder = &decimalFloatBuilder{}
 
-func generateDecimalFloatBuilder(ctx *Context) ObjectBuilder { return globalDecimalFloatBuilder }
+func generateDecimalFloatBuilder(ctx *Context) Builder { return globalDecimalFloatBuilder }
 func (_this *decimalFloatBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *decimalFloatBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
@@ -186,7 +186,7 @@ type floatBuilder struct{}
 
 var globalFloatBuilder = &floatBuilder{}
 
-func generateFloatBuilder(ctx *Context) ObjectBuilder { return globalFloatBuilder }
+func generateFloatBuilder(ctx *Context) Builder { return globalFloatBuilder }
 func (_this *floatBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *floatBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
@@ -224,7 +224,7 @@ type intBuilder struct{}
 
 var globalIntBuilder = &intBuilder{}
 
-func generateIntBuilder(ctx *Context) ObjectBuilder { return globalIntBuilder }
+func generateIntBuilder(ctx *Context) Builder { return globalIntBuilder }
 func (_this *intBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *intBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
@@ -262,7 +262,7 @@ type pBigDecimalFloatBuilder struct{}
 
 var globalPBigDecimalFloatBuilder = &pBigDecimalFloatBuilder{}
 
-func generatePBigDecimalFloatBuilder(ctx *Context) ObjectBuilder { return globalPBigDecimalFloatBuilder }
+func generatePBigDecimalFloatBuilder(ctx *Context) Builder { return globalPBigDecimalFloatBuilder }
 func (_this *pBigDecimalFloatBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *pBigDecimalFloatBuilder) BuildFromNil(ctx *Context, dst reflect.Value) reflect.Value {
@@ -305,7 +305,7 @@ type pBigFloatBuilder struct{}
 
 var globalPBigFloatBuilder = &pBigFloatBuilder{}
 
-func generatePBigFloatBuilder(ctx *Context) ObjectBuilder { return globalPBigFloatBuilder }
+func generatePBigFloatBuilder(ctx *Context) Builder { return globalPBigFloatBuilder }
 func (_this *pBigFloatBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *pBigFloatBuilder) BuildFromNil(ctx *Context, dst reflect.Value) reflect.Value {
@@ -348,7 +348,7 @@ type pBigIntBuilder struct{}
 
 var globalPBigIntBuilder = &pBigIntBuilder{}
 
-func generatePBigIntBuilder(ctx *Context) ObjectBuilder { return globalPBigIntBuilder }
+func generatePBigIntBuilder(ctx *Context) Builder { return globalPBigIntBuilder }
 func (_this *pBigIntBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *pBigIntBuilder) BuildFromNil(ctx *Context, dst reflect.Value) reflect.Value {
@@ -394,7 +394,7 @@ type uintBuilder struct {
 
 var globalUintBuilder = &uintBuilder{}
 
-func generateUintBuilder(ctx *Context) ObjectBuilder { return globalUintBuilder }
+func generateUintBuilder(ctx *Context) Builder { return globalUintBuilder }
 func (_this *uintBuilder) String() string            { return reflect.TypeOf(_this).String() }
 
 func (_this *uintBuilder) BuildFromInt(ctx *Context, value int64, dst reflect.Value) reflect.Value {
