+++ rules/rules_event_rcv.go
@@ -1,0 +1,309 @@
+// Copyright 2019 Karl Stenerud
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to
+// deal in the Software without restriction, including without limitation the
+// rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+// sell copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+// IN THE SOFTWARE.
+
+package rules
+
+import (
+	"math"
+	"math/big"
+	"time"
+
+	"github.com/kstenerud/go-concise-encoding/events"
+	"github.com/kstenerud/go-concise-encoding/internal/common"
+	"github.com/kstenerud/go-concise-encoding/options"
+	"github.com/kstenerud/go-concise-encoding/version"
+
+	"github.com/cockroachdb/apd/v2"
+	"github.com/kstenerud/go-compact-float"
+	"github.com/kstenerud/go-compact-time"
+)
+
+// RulesEventReceiver is a DataEventsReceiver passthrough object that constrains
+// the order and contents of events to ensure that they form a valid and
+// complete Concise Encoding document.
+//
+// Put this right after your event generator in the event receiver chain to
+// enforce correctly formed documents.
+//
+// Note: This is a LOW LEVEL API. Error reporting is done via panics. Be sure
+// to recover() at an appropriate location when calling this struct's methods
+// directly (with the exception of constructors and initializers, which are not
+// designed to panic).
+type RulesEventReceiver struct {
+	context  Context
+	receiver events.DataEventReceiver
+}
+
+// Create a new rules set.
+// If opts = nil, defaults are used.
+func NewRules(nextReceiver events.DataEventReceiver, opts *options.RuleOptions) *RulesEventReceiver {
+	_this := &RulesEventReceiver{}
+	_this.Init(nextReceiver, opts)
+	return _this
+}
+
+// Initialize a rules set.
+// If opts = nil, defaults are used.
+func (_this *RulesEventReceiver) Init(nextReceiver events.DataEventReceiver, opts *options.RuleOptions) {
+	opts = opts.WithDefaultsApplied()
+	_this.receiver = nextReceiver
+	_this.context.Init(version.ConciseEncodingVersion, opts)
+}
+
+// Reset the rules set back to its initial state.
+func (_this *RulesEventReceiver) Reset() {
+	_this.context.Reset()
+}
+
+func (_this *RulesEventReceiver) SetNextReceiver(nextReceiver events.DataEventReceiver) {
+	_this.receiver = nextReceiver
+}
+
+func (_this *RulesEventReceiver) OnBeginDocument() {
+	_this.context.CurrentEntry.Rule.OnBeginDocument(&_this.context)
+	_this.receiver.OnBeginDocument()
+}
+
+func (_this *RulesEventReceiver) OnVersion(version uint64) {
+	_this.context.CurrentEntry.Rule.OnVersion(&_this.context, version)
+	_this.receiver.OnVersion(version)
+}
+
+func (_this *RulesEventReceiver) OnPadding(count int) {
+	_this.context.CurrentEntry.Rule.OnPadding(&_this.context)
+	_this.receiver.OnPadding(count)
+}
+
+func (_this *RulesEventReceiver) OnNA() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnNA(&_this.context)
+	_this.receiver.OnNA()
+}
+
+func (_this *RulesEventReceiver) OnBool(value bool) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnBool(value)
+}
+
+func (_this *RulesEventReceiver) OnTrue() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnTrue()
+}
+
+func (_this *RulesEventReceiver) OnFalse() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnFalse()
+}
+
+func (_this *RulesEventReceiver) OnPositiveInt(value uint64) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnPositiveInt(&_this.context, value)
+	_this.receiver.OnPositiveInt(value)
+}
+
+func (_this *RulesEventReceiver) OnNegativeInt(value uint64) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnNegativeInt(value)
+}
+
+func (_this *RulesEventReceiver) OnInt(value int64) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnInt(&_this.context, value)
+	_this.receiver.OnInt(value)
+}
+
+func (_this *RulesEventReceiver) OnBigInt(value *big.Int) {
+	if value == nil {
+		_this.OnNA()
+		return
+	}
+
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnBigInt(&_this.context, value)
+	_this.receiver.OnBigInt(value)
+}
+
+func (_this *RulesEventReceiver) OnFloat(value float64) {
+	if math.IsNaN(value) {
+		_this.OnNan(common.IsSignalingNan(value))
+		return
+	}
+
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnFloat(&_this.context, value)
+	_this.receiver.OnFloat(value)
+}
+
+func (_this *RulesEventReceiver) OnBigFloat(value *big.Float) {
+	if value == nil {
+		_this.OnNA()
+		return
+	}
+
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnBigFloat(&_this.context, value)
+	_this.receiver.OnBigFloat(value)
+}
+
+func (_this *RulesEventReceiver) OnDecimalFloat(value compact_float.DFloat) {
+	if value.IsNan() {
+		_this.OnNan(value.IsSignalingNan())
+		return
+	}
+
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnDecimalFloat(&_this.context, value)
+	_this.receiver.OnDecimalFloat(value)
+}
+
+func (_this *RulesEventReceiver) OnBigDecimalFloat(value *apd.Decimal) {
+	if value == nil {
+		_this.OnNA()
+		return
+	}
+	if value.Form == apd.NaNSignaling {
+		_this.OnNan(true)
+		return
+	}
+	if value.Form == apd.NaN {
+		_this.OnNan(false)
+		return
+	}
+
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnBigDecimalFloat(&_this.context, value)
+	_this.receiver.OnBigDecimalFloat(value)
+}
+
+func (_this *RulesEventReceiver) OnNan(signaling bool) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnNonKeyableObject(&_this.context)
+	_this.receiver.OnNan(signaling)
+}
+
+func (_this *RulesEventReceiver) OnUUID(value []byte) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnUUID(value)
+}
+
+func (_this *RulesEventReceiver) OnTime(value time.Time) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnTime(value)
+}
+
+func (_this *RulesEventReceiver) OnCompactTime(value compact_time.Time) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnKeyableObject(&_this.context)
+	_this.receiver.OnCompactTime(value)
+}
+
+func (_this *RulesEventReceiver) OnArray(arrayType events.ArrayType, elementCount uint64, value []byte) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnArray(&_this.context, arrayType, elementCount, value)
+	_this.receiver.OnArray(arrayType, elementCount, value)
+}
+
+func (_this *RulesEventReceiver) OnStringlikeArray(arrayType events.ArrayType, value string) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnStringlikeArray(&_this.context, arrayType, value)
+	_this.receiver.OnStringlikeArray(arrayType, value)
+}
+
+func (_this *RulesEventReceiver) OnArrayBegin(arrayType events.ArrayType) {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnArrayBegin(&_this.context, arrayType)
+	_this.receiver.OnArrayBegin(arrayType)
+}
+
+func (_this *RulesEventReceiver) OnArrayChunk(length uint64, moreChunksFollow bool) {
+	_this.context.CurrentEntry.Rule.OnArrayChunk(&_this.context, length, moreChunksFollow)
+	_this.receiver.OnArrayChunk(length, moreChunksFollow)
+}
+
+func (_this *RulesEventReceiver) OnArrayData(data []byte) {
+	_this.context.CurrentEntry.Rule.OnArrayData(&_this.context, data)
+	_this.receiver.OnArrayData(data)
+}
+
+func (_this *RulesEventReceiver) OnConcatenate() {
+	panic("TODO: Rules.OnConcatenate")
+}
+
+func (_this *RulesEventReceiver) OnList() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnList(&_this.context)
+	_this.receiver.OnList()
+}
+
+func (_this *RulesEventReceiver) OnMap() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnMap(&_this.context)
+	_this.receiver.OnMap()
+}
+
+func (_this *RulesEventReceiver) OnMarkup() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnMarkup(&_this.context)
+	_this.receiver.OnMarkup()
+}
+
+func (_this *RulesEventReceiver) OnMetadata() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnMetadata(&_this.context)
+	_this.receiver.OnMetadata()
+}
+
+func (_this *RulesEventReceiver) OnComment() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnComment(&_this.context)
+	_this.receiver.OnComment()
+}
+
+func (_this *RulesEventReceiver) OnEnd() {
+	_this.context.CurrentEntry.Rule.OnEnd(&_this.context)
+	_this.receiver.OnEnd()
+}
+
+func (_this *RulesEventReceiver) OnMarker() {
+	_this.context.CurrentEntry.Rule.OnMarker(&_this.context)
+	_this.receiver.OnMarker()
+}
+
+func (_this *RulesEventReceiver) OnReference() {
+	_this.context.NotifyNewObject()
+	_this.context.CurrentEntry.Rule.OnReference(&_this.context)
+	_this.receiver.OnReference()
+}
+
+func (_this *RulesEventReceiver) OnConstant(name []byte, explicitValue bool) {
+	_this.context.CurrentEntry.Rule.OnConstant(&_this.context, name, explicitValue)
+	_this.receiver.OnConstant(name, explicitValue)
+}
+
+func (_this *RulesEventReceiver) OnEndDocument() {
+	_this.context.CurrentEntry.Rule.OnEndDocument(&_this.context)
+	_this.receiver.OnEndDocument()
+}
